requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.0.1
django==4.2.29
django-unfold==0.67.0
gunicorn==23.0.0
idna==3.11
importlib-metadata==8.7.1
markdown==3.9
packaging==26.0
pillow==11.3.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
zipp==3.23.0
